ViewModel Validaciones Ok
